+++ packages/math-input/src/components/keypad/__tests__/keypad.test.tsx
@@ -4,7 +4,7 @@ import "@testing-library/jest-dom";
 
 import keyConfigs from "../../../data/key-configs";
 import {CursorContext} from "../../input/cursor-contexts";
-import Keypad from "../index";
+import Keypad from "../keypad";
 
 const contextToKeyAria = {
     [CursorContext.IN_PARENS]: keyConfigs.JUMP_OUT_PARENTHESES.ariaLabel,
