+++ testing/render-keypad-with-cypress.tsx
@@ -1,7 +1,7 @@
 import {mount} from "@cypress/react";
 import * as React from "react";
 
-import Keypad from "../packages/math-input/src/components/keypad";
+import Keypad from "../packages/math-input/src/components/keypad/keypad";
 
 const renderSingleKeypad = (handleClickKey) =>
     mount(
